+++ web/client/.eslintignore
@@ -1,2 +1,2 @@
 src/pages/Rollout.js
-src/pages/MinifiedSunwell.js
+src/sunwell/MinifiedSunwell.js
